# utilities/openapi.yaml
openapi: 3.0.0
info:
  title: "Mozilla IAM"
  description: IAM is a service to match authorization requests against predefined policies.
  version: "0.1"
  contact:
    url: "irc://irc.mozilla.org:6696/#product-delivery"
  license:
    name: "Mozilla Public License 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
...
paths:
...
  /allowed:
    post:
      summary: "Check if the specified request is allowed"
      operationId: "allowed"
      produces:
      - "application/json"
      requestBody:
        description: Authorization request
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                principals:
                  type: array
                  items:
                    type: string
                action:
                  type: string
                resource:
                  type: string
                context:
                  type: object
      responses:
        "400":
          description: "Invalid posted data."
        "401":
          description: "Authentication failed."
        "403":
          description: "JWT token is invalid (authorization failed)."
        "200":
          description: "Return whether it is allowed or not."
          schema:
            type: object
            properties:
              allowed:
                type: boolean
              principals:
                type: array
                items:
                  type: string
      tags:
      - Doorman
...